backend/apps/accounts/urls.py
"""
URL patterns for the accounts app.
"""
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenRefreshView

from .views import (
    RegisterView,
    LoginView,
    LogoutView,
    MeView,
    PasswordChangeView,
    UserViewSet,
    AgentViewSet,
    CustomerViewSet,
    TeamViewSet,
)

router = DefaultRouter()
router.register(r"users", UserViewSet, basename="user")
router.register(r"agents", AgentViewSet, basename="agent")
router.register(r"customers", CustomerViewSet, basename="customer")
router.register(r"teams", TeamViewSet, basename="team")

urlpatterns = [
    path("register/", RegisterView.as_view(), name="register"),
    path("login/", LoginView.as_view(), name="login"),
    path("logout/", LogoutView.as_view(), name="logout"),
    path("refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("me/", MeView.as_view(), name="me"),
    path("password/change/", PasswordChangeView.as_view(), name="password_change"),
    path("", include(router.urls)),
]
