backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-extensions==3.2.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-storages==1.14.3
django-redis==5.4.0

channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

celery==5.4.0
redis==5.0.4

psycopg2-binary==2.9.9
dj-database-url==2.2.0

Pillow==10.3.0
python-dotenv==1.0.1
python-dateutil==2.9.0
whitenoise==6.6.0
gunicorn==22.0.0

boto3==1.34.69
bleach==6.1.0
markdown==3.6
